+++ OpenMarket/OpenMarket/Error/JSONError.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum JSONError: LocalizedError {
+    case dataDecodeFailed
+}
