+++ OpenMarket/OpenMarket/Error/URLSessionError.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum URLSessionError: LocalizedError {
+    case requestFailed
+    case responseFailed(code: Int)
+    case invaildData
+}
