+++ OpenMarket/OpenMarket/Model/JSONModel/Image.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+struct Image: Codable {
+    let id: Int
+    let url, thumbnailURL: String
+    let succeed: Bool
+    let issuedAt: String
+
+    enum CodingKeys: String, CodingKey {
+        case id, url
+        case thumbnailURL = "thumbnail_url"
+        case succeed
+        case issuedAt = "issued_at"
+    }
+}
