+++ OpenMarket/OpenMarket/Model/JSONModel/Vendors.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+struct Vendors: Codable {
+    let name: String
+    let id: Int
+    let createdAt, issuedAt: String
+
+    enum CodingKeys: String, CodingKey {
+        case name, id
+        case createdAt = "created_at"
+        case issuedAt = "issued_at"
+    }
+}
